Refactor Mains to use common BitcoinSRunner
