+++ app/server/src/main/scala/org/bitcoins/server/BitcoindRpcAppConfig.scala
@@ -107,7 +107,7 @@ case class BitcoindRpcAppConfig(
   lazy val client: BitcoindRpcClient = {
     val version = bitcoindInstance.getVersion
     implicit val system: ActorSystem =
-      ActorSystem.create("bitcoind-rpc-client-created-by-bitcoin-s")
+      ActorSystem.create("bitcoind-rpc-client-created-by-bitcoin-s", config)
     BitcoindRpcClient.fromVersion(version, bitcoindInstance)
   }
